Prevent Scope from triggering a cache reload
When a file gets changed in watch mode or after a fix, all the files get
re-reviewed, but the errors and context in an internal cache are used if
neither the module's AST nor the imported modules' context have changed.

Scope was storing the signatures of all the modules in a module's
associated context. That means that if a module's signature changed (new
function), then all modules that imported it directly or indirectly would
be recomputed since every context would be invalidated.

Now, instead, only the modules that import the updated module directly
get recomputed, and the ones that import it indirectly only get recomputed if
the intermediate module's context changes due to the re-analysis.

This should not have any impact on the results you get from Scope. The
only differences are a speed bump on rewatch and potentially a memory
usage reduction.